Bug fix, and fixes of warning and santizer test.
STATS_CHANGED

+++ av1/common/reconinter.h
@@ -388,7 +388,7 @@ void av1_build_inter_predictors(const AV1_COMMON *cm, MACROBLOCKD *xd,
                                 const BUFFER_SET *dst_orig,
 #endif
                                 int build_for_obmc, int bw, int bh, int mi_x,
-                                int mi_y, uint8_t **mc_buf,
+                                int mi_y, uint16_t **mc_buf,
                                 CalcSubpelParamsFunc calc_subpel_params_func);
 
 #if CONFIG_OPTFLOW_REFINEMENT
